+++ src/main/java/dariusG82/custom_exeptions/OrderDoesNotExistException.java
@@ -2,7 +2,7 @@ package dariusG82.custom_exeptions;
 
 public class OrderDoesNotExistException extends Exception {
 
-    public OrderDoesNotExistException(int number) {
+    public OrderDoesNotExistException(long number) {
         super("Order " + number + " doesn't exist");
     }
 }
